ref: move view routes to tms_datashader_api/views.py
